fix(gameplay): CodeRabbit review findings on the collapse solver (#786)
Six findings, all verified against the code before acting, all valid.

  * The graph's staleness signature folded only entity ids, but Rebuild also
    caches m_Anchor, m_ContactMargin and m_MaxLateralSpan. Lua, the editor
    inspector, MCP and C# can all write those mid-Play and none of them know the
    graph exists, so retuning a piece left the solver answering with the values
    it used to have. The signature now covers every authored input the build
    reads. m_State stays out of it on purpose — a piece leaving is tracked in
    place, and folding state in would rebuild every tick of a collapse. Pinned by
    RetuningAPieceMidPlayInvalidatesTheGraph, which also asserts that an
    unchanged scene does NOT rebuild.

  * Median-sized hash cells only changed the constant on the O(N^2) cliff they
    were meant to remove: a query has to reach the largest partner it could
    overlap, so one oversized piece still widened every other piece's query.
    Oversized pieces are now excluded from the grid pass and paired by a linear
    scan, with the grid bounded by the largest NORMAL radius. The pair test moved
    into one shared lambda so both passes cannot drift. The stale comment
    claiming every query sweeps at most 3x3x3 cells is gone.

  * The Dynamic branch of CreateBodySettings passed m_Mass to CalculateInertia
    unvalidated while the adjacent branch guarded it — a zero or NaN mass is a
    Jolt assert, and nothing validates the field on the way in. Both branches now
    share one guard, and an invalid mass is reported rather than silently
    substituted.

  * The save-game restore computed its timer ceiling as m_CollapseHops + 1 in
    u32, so a corrupt UINT32_MAX wrapped to zero and detached the piece instantly
    — the opposite of the conservative reading of an untrusted value. Widened
    before the add.

  * DestructibleComponent's inspector exposed every field except m_ChunkMesh, so
    the pre-authored debris asset (the intended production path in ADR 0013) was
    the one thing still unreachable from the editor. Added as a MeshSource-typed
    drop slot, following the Weather Map / Tileset idiom.

  * ADR 0021 and the agent-rules guide still described invalidation as a piece
    COUNT, which stopped being true two commits ago. Both now state the signature
    and its two consequences outright: a same-count create/destroy swap rebuilds,
    unchanged identities and inputs do not — and the one input deliberately left
    uncovered (a standing piece's transform) is named rather than left implied.

22/22 destruction tests, 286/286 across physics/joint/vehicle/ragdoll/cloth/
character/save-game/coverage.

+++ OloEngine/src/OloEngine/SaveGame/SaveGameComponentSerializer.cpp
@@ -1620,8 +1620,12 @@ namespace OloEngine
             // state it is in — the Detaching delay is m_CollapseDelay per hop, so
             // it has no fixed ceiling of its own and a flat clamp would quietly
             // rewrite the collapse ordering this field exists to preserve.
+            // Widen before the +1: a corrupt UINT32_MAX hop count would wrap to
+            // zero, collapse the ceiling to zero, and drop the timer to zero —
+            // detaching the piece instantly, which is the opposite of the
+            // conservative thing to do with an untrusted value.
             const f32 ceiling = (c.m_State == StructuralState::Detaching)
-                                    ? c.m_CollapseDelay * static_cast<f32>(c.m_CollapseHops + 1u)
+                                    ? c.m_CollapseDelay * static_cast<f32>(static_cast<u64>(c.m_CollapseHops) + 1ull)
                                     : ((c.m_State == StructuralState::Falling) ? c.m_FallDuration : 0.0f);
             c.m_StateTimer = (!std::isfinite(c.m_StateTimer) || c.m_StateTimer < 0.0f) ? 0.0f
                                                                                        : std::min(c.m_StateTimer, ceiling);
